move OPGroup tests - from test_group_get.py, to - test_op_objec_types/test_op_group.py
